+++ packages/graphql-codegen-cli/src/codegen.ts
@@ -18,16 +18,10 @@ import { NoTypeDefinitionsFound, type UnnormalizedTypeDefPointer } from '@graphq
 import { mergeTypeDefs } from '@graphql-tools/merge';
 import { CodegenContext, ensureContext } from './config.js';
 import { getDocumentTransform } from './documentTransforms.js';
+import { isESMModule } from './isESMModule.js';
 import { getPluginByName } from './plugins.js';
 import { getPresetByName } from './presets.js';
 import { debugLog, printLogs } from './utils/debugging.js';
-
-/**
- * Poor mans ESM detection.
- * Looking at this and you have a better method?
- * Send a PR.
- */
-const isESMModule = (typeof __dirname === 'string') === false;
 
 const makeDefaultLoader = (from: string) => {
   if (fs.statSync(from).isDirectory()) {
@@ -44,6 +38,8 @@ const makeDefaultLoader = (from: string) => {
            * as import.meta is unavailable in a CommonJS context
            * and furthermore unavailable in stable Node.js.
            **/
+          // FIXME(pnpm-update): this causes dev-test devDeps to be brought into CLI's package.json, which is not ideal.
+          // Note that `relativeRequire.resolve(mod)` seems to work correctly for ESM as well.
           mod
         : relativeRequire.resolve(mod)
     );